+++ src/SlmQueue/Queue/ProvidesQueue.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace SlmQueue\Queue;
-
-trait ProvidesQueue
-{
-    use QueueAwareTrait;
-}